Settings: Do not allow to access quick settings on secure lockscreen [2/2]
We have a lot security guys as user due of using the phones as business phones...that
said a lot complains go into the direction of google how it is possible to have
access to quick settings on a secure lockscreen which is absolutely understandable
due that you are able to control Wifi BT AirplaneMode data etc.

This patch takes care on this problem and adds an option to control this behaviour.

ATTENTION: Due that this is really an UX derp for an average user this feature is
enabled by default now. User need to disable it explicit in settings to get on secure
locksreen the original android behaviour back.

Thanks to @chickenfactory got this commit: https://github.com/temasek/android_packages_apps_Settings/commit/aa09077d2eac4ad91def32fc2b5d86cd2841ae5d

Change-Id: I700d5283035174ec679a08539b5dc85e137b9578

Conflicts:
	res/xml/notification_drawer_settings.xml

Conflicts:
	res/values/cr_strings.xml